created homepage, removed code from app file, and linked new home to APP file.
