+++ client/package.json
@@ -3,6 +3,7 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "@apollo/client": "^3.6.9",
     "@arwes/animation": "^1.0.0-alpha.19",
     "@arwes/core": "^1.0.0-alpha.19",
     "@arwes/design": "^1.0.0-alpha.19",
@@ -17,10 +18,12 @@
     "@testing-library/user-event": "^13.5.0",
     "animejs": "^3.2.1",
     "framer-motion": "^7.2.1",
+    "graphql": "^16.6.0",
     "howler": "2.2",
     "polished": "^4.2.2",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-router-dom": "^6.4.0",
     "react-scripts": "5.0.1",
     "web-vitals": "^2.1.4"
   },
